Input: add support for ChipOne icn8505 based touchscreens
The ChipOne icn8505 is an i2c capacitive touchscreen controller typically
used in cheap x86 tablets, this commit adds a driver for it.

Note the icn8505 is somewhat similar to the icn8318 and I started with
modifying that driver to support both, but in the end the differences were
too large and I decided to write a new driver instead.

Cc: Antonio Davide Trogu <trogu.davide@gmail.com>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>
---
Changes in v2:
-Make icn8505_write/read_xfer return 0 rather then a positive value on success
-Rename the "ret" local variable to "error" where applicable
-Move printing of attempt %d/%d on firmware upload issues from
 icn8505_try_fw_upload() to its caller, icn8505_upload_fw()
-Drop inline marking from icn8505_touch_active()
-Drop duplicate input->dev.parent = dev;
-Use __maybe_unused for suspend/resume() instead of #ifdef CONFIG_PM_SLEEP